+++ ci4/app/Libraries/DeploymentSteps/MigrationJobStep.php
@@ -286,6 +286,9 @@ class MigrationJobStep extends BaseDeploymentStep {
         if (strlen($containerImage->security_context_run_as_group) > 0) {
             $container->setAttribute('securityContext.runAsGroup', (int)$containerImage->security_context_run_as_group);
         }
+        if (strlen($containerImage->security_context_fs_group) > 0) {
+            $container->setAttribute('securityContext.fsGroup', (int)$containerImage->security_context_fs_group);
+        }
         $container->setAttribute('securityContext.allowPrivilegeEscalation', (bool)$containerImage->security_context_allow_privilege_escalation);
         $container->setAttribute('securityContext.readOnlyRootFilesystem', (bool)$containerImage->security_context_read_only_root_filesystem);