+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v1.5.5 (2025-06-19)
+
+### Fixed bugs
+* 
+
+### Enhancements
+* Added `fsGroup` to security context
+
+### Upgrade guide
+1. Deploy new image
+2. Migrate database [(Guide)](https://github.com/4spacesdk/kubernetes-service-orchestrator?tab=readme-ov-file#migrate-database-helm)
+
+
+
 ## v1.5.4 (2025-06-15)
 
 ### Fixed bugs